# server/gen/tasks/model_approver_response.go
/*
 * Tasks Hub
 *
 * This is a sample Tasks Hub Server based on the OpenAPI 3.0 specification.  Some useful links: - [The Tasks Hub repository](https://github.com/vbelouso/tasks) - [The source API definition](https://github.com/vbelouso/tasks/api/openapi-spec/v1/openapi.yaml)
 *
 * API version: 1.0.11
 * Contact: fake@msbang.io
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package tasks

type ApproverResponse struct {

	Id int64 `json:"id"`

	Name string `json:"name"`

	Email string `json:"email"`
}

// AssertApproverResponseRequired checks if the required fields are not zero-ed
func AssertApproverResponseRequired(obj ApproverResponse) error {
	elements := map[string]interface{}{
		"id": obj.Id,
		"name": obj.Name,
		"email": obj.Email,
	}
	for name, el := range elements {
		if isZero := IsZeroValue(el); isZero {
			return &RequiredError{Field: name}
		}
	}

	return nil
}

// AssertRecurseApproverResponseRequired recursively checks if required fields are not zero-ed in a nested slice.
// Accepts only nested slice of ApproverResponse (e.g. [][]ApproverResponse), otherwise ErrTypeAssertionError is thrown.
func AssertRecurseApproverResponseRequired(objSlice interface{}) error {
	return AssertRecurseInterfaceRequired(objSlice, func(obj interface{}) error {
		aApproverResponse, ok := obj.(ApproverResponse)
		if !ok {
			return ErrTypeAssertionError
		}
		return AssertApproverResponseRequired(aApproverResponse)
	})
}