# server/gen/tasks/model_task_request.go
/*
 * Tasks Hub
 *
 * This is a sample Tasks Hub Server based on the OpenAPI 3.0 specification.  Some useful links: - [The Tasks Hub repository](https://github.com/vbelouso/tasks) - [The source API definition](https://github.com/vbelouso/tasks/api/openapi-spec/v1/openapi.yaml)
 *
 * API version: 1.0.11
 * Contact: fake@msbang.io
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package tasks

type TaskRequest struct {

	Title string `json:"title"`

	Description string `json:"description"`

	Approvers []ApproverRequest `json:"approvers,omitempty"`
}

// AssertTaskRequestRequired checks if the required fields are not zero-ed
func AssertTaskRequestRequired(obj TaskRequest) error {
	elements := map[string]interface{}{
		"title": obj.Title,
		"description": obj.Description,
	}
	for name, el := range elements {
		if isZero := IsZeroValue(el); isZero {
			return &RequiredError{Field: name}
		}
	}

	for _, el := range obj.Approvers {
		if err := AssertApproverRequestRequired(el); err != nil {
			return err
		}
	}
	return nil
}

// AssertRecurseTaskRequestRequired recursively checks if required fields are not zero-ed in a nested slice.
// Accepts only nested slice of TaskRequest (e.g. [][]TaskRequest), otherwise ErrTypeAssertionError is thrown.
func AssertRecurseTaskRequestRequired(objSlice interface{}) error {
	return AssertRecurseInterfaceRequired(objSlice, func(obj interface{}) error {
		aTaskRequest, ok := obj.(TaskRequest)
		if !ok {
			return ErrTypeAssertionError
		}
		return AssertTaskRequestRequired(aTaskRequest)
	})
}